Fix starting ores for MediumPy.
PyRO introduces tin for science buildings and quarts for the science packs.
Both should therefore be starting ores. You also need aluminium for the car,
so it should also be a starting ore.

PyAL makes planter boxes for automation science, and you can get the car
with them, so tin and quartz no longer need to be starting ores.
Aluminium still needs to be a starting ore, as it's still needed for the car.

RSO is also misconfigured with respect to these starting ores.
I guess I'll open a ticket on their page to resolve this or something.

Resolves pyanodon/pybugreports#133